+++ src/pages/hyungwon/Main/DefaultComment/DefaultComment.scss
@@ -1,0 +1,30 @@
+.add_comment {
+  display: flex;
+  justify-content: space-between;
+  height: auto;
+  .comment_right {
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+    margin: 0px 0px 0px 15px;
+    font-size: 12px;
+    .nickname {
+      font-weight: 900;
+      margin: 0px 10px 0px 10px;
+    }
+  }
+  .comment_button {
+    .delete_comment {
+      margin-right: 10px;
+      width: 20px;
+      height: 20px;
+      cursor: pointer;
+    }
+    .comment_heart {
+      width: 20px;
+      height: 20px;
+      margin-right: 15px;
+      cursor: pointer;
+    }
+  }
+}
